+++ packages/devtools_app_shared/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add `FlutterEvent` and `DeveloperServiceEvent` constants.
 * Add `connectedAppPackageRoot`, `rootPackageDirectoryForMainIsolate`, and
 `mainIsolateRootLibraryUriAsString` methods to the `ServiceManager` class.
+* Bump minimum Dart SDK version to Dart stable `3.4.3` and minimum Flutter SDK
+version to Flutter stable `3.22.2`.
 
 ## 0.2.0-dev.0
 * Add `tooltipWaitExtraLong` to `utils.dart`.